+++ cmd/earthly/registry_cmds.go
@@ -1,13 +1,17 @@
 package main
 
 import (
+	"encoding/json"
 	"fmt"
 	"io"
 	"os"
+	"path"
+	"sort"
 	"strings"
 	"text/tabwriter"
 
 	"github.com/earthly/earthly/cloud"
+	"github.com/moby/buildkit/session/secrets"
 	"github.com/pkg/errors"
 	"github.com/urfave/cli/v2"
 )
@@ -15,12 +19,21 @@ import (
 func (app *earthlyApp) registryCmds() []*cli.Command {
 	return []*cli.Command{
 		{
-			Name:        "login",
-			Usage:       "Login and store credentials in earthly-cloud *beta*",
-			Description: "Login and store credentials in earthly-cloud *beta*",
-			UsageText: "earthly registry login --username <username> --password <password> [<host>]\n" +
-				"	earthly registry login --org <org> --project <project> --username <username> --password <password> [<host>]\n",
-			Action: app.actionRegistryLogin,
+			Name:        "setup",
+			Usage:       "setup, and store, registry credentials in the earthly-cloud *beta*",
+			Description: "setup, and store, registry credentials in the earthly-cloud *beta*",
+			UsageText: "earthly registry setup [--org <org> --project <project>] [--cred-helper <none|ecr-login|gcloud>] ...\n\n" +
+				"username/password based registry (--cred-helper=none):\n" +
+				"	earthly registry setup --username <username> --password <password> [<host>]\n" +
+				"	earthly registry setup --org <org> --project <project> --username <username> --password <password> [<host>]\n\n" +
+				"AWS elastic container registry (--cred-helper=ecr-login):\n" +
+				"	earthly registry setup --cred-helper ecr-login --aws-access-key-id <key> --aws-secret-access-key <secret> <host>\n" +
+				"	earthly registry setup --org <org> --project <project> --cred-helper ecr-login --aws-access-key-id <key> --aws-secret-access-key <secret> <host>\n\n" +
+				"GCP artifact or container registry (--cred-helper=gcloud):\n" +
+				"	earthly registry setup --cred-helper gcloud --gcp-key <key> <host>\n" +
+				"	earthly registry setup --org <org> --project <project> --cred-helper gcloud --gcp-key <key> <host>\n" +
+				"",
+			Action: app.actionRegistrySetup,
 			Flags: []cli.Flag{
 				&cli.StringFlag{
 					Name:        "org",
@@ -35,10 +48,17 @@ func (app *earthlyApp) registryCmds() []*cli.Command {
 					Destination: &app.projectName,
 				},
 				&cli.StringFlag{
+					Name:        "cred-helper",
+					EnvVars:     []string{"EARTHLY_REGISTRY_CRED_HELPER"},
+					Usage:       "Use a credential helper when logging into the registry (ecr-login, gcloud).",
+					Required:    false,
+					Destination: &app.registryCredHelper,
+				},
+				&cli.StringFlag{
 					Name:        "username",
 					EnvVars:     []string{"EARTHLY_REGISTRY_USERNAME"},
 					Usage:       "The username to use when logging into the registry.",
-					Required:    true,
+					Required:    false,
 					Destination: &app.registryUsername,
 				},
 				&cli.StringFlag{
@@ -54,6 +74,27 @@ func (app *earthlyApp) registryCmds() []*cli.Command {
 					Usage:       "Read the password from stdin (recommended).",
 					Required:    false,
 					Destination: &app.registryPasswordStdin,
+				},
+				&cli.StringFlag{
+					Name:        "aws-access-key-id",
+					EnvVars:     []string{"AWS_ACCESS_KEY_ID"},
+					Usage:       "AWS access key ID to use for elastic-container-registry.",
+					Required:    false,
+					Destination: &app.awsAccessKeyID,
+				},
+				&cli.StringFlag{
+					Name:        "aws-secret-access-key",
+					EnvVars:     []string{"AWS_SECRET_ACCESS_KEY"},
+					Usage:       "AWS secret access key to use for elastic-container-registry.",
+					Required:    false,
+					Destination: &app.awsSecretAccessKey,
+				},
+				&cli.StringFlag{
+					Name:        "gcp-key",
+					EnvVars:     []string{"GCP_KEY"},
+					Usage:       "GCP key to use for artifact or container registry.",
+					Required:    false,
+					Destination: &app.gcpKey,
 				},
 			},
 		},
@@ -79,11 +120,11 @@ func (app *earthlyApp) registryCmds() []*cli.Command {
 			},
 		},
 		{
-			Name:  "logout",
-			Usage: "Logout of a registry (that has credentials stored in earthly-cloud) *beta*",
-			UsageText: "earthly registry logout [<host>]\n" +
-				"	earthly registry login --org <org> --project <project> [<host>]\n",
-			Action: app.actionRegistryLogout,
+			Name:  "remove",
+			Usage: "Remove stored registry credentials *beta*",
+			UsageText: "earthly registry remove [<host>]\n" +
+				"	earthly registry remove [--org <org> --project <project>] <host>\n",
+			Action: app.actionRegistryRemove,
 			Flags: []cli.Flag{
 				&cli.StringFlag{
 					Name:        "org",
@@ -126,8 +167,8 @@ func (app *earthlyApp) getRegistriesPath() (string, error) {
 	return fmt.Sprintf("/%s/%s/std/registry/", app.orgName, app.projectName), nil
 }
 
-func (app *earthlyApp) actionRegistryLogin(cliCtx *cli.Context) error {
-	app.commandName = "registryLogin"
+func (app *earthlyApp) actionRegistrySetup(cliCtx *cli.Context) error {
+	app.commandName = "registrySetup"
 
 	path, err := app.getRegistriesPath()
 	if err != nil {
@@ -152,6 +193,53 @@ func (app *earthlyApp) actionRegistryLogin(cliCtx *cli.Context) error {
 		return fmt.Errorf("hosts is malformed")
 	}
 
+	switch app.registryCredHelper {
+	case "", "none":
+		return app.actionRegistrySetupUsernamePassword(cliCtx, path, cloudClient, host)
+	case "ecr-login":
+		return app.actionRegistrySetupECRLogin(cliCtx, path, cloudClient, host)
+	case "gcloud":
+		return app.actionRegistrySetupGCloud(cliCtx, path, cloudClient, host)
+	default:
+		return fmt.Errorf("unsupported credential helper %s", app.registryCredHelper)
+	}
+}
+
+func (app *earthlyApp) actionRegistrySetupECRLogin(cliCtx *cli.Context, path string, cloudClient *cloud.Client, host string) error {
+	if app.awsAccessKeyID == "" {
+		return fmt.Errorf("--aws-access-key-id is missing (or empty)")
+	}
+	if app.awsSecretAccessKey == "" {
+		return fmt.Errorf("--aws-secret-access-key is missing (or empty)")
+	}
+	err := cloudClient.SetSecret(cliCtx.Context, path+host+"/cred_helper", []byte("ecr-login"))
+	if err != nil {
+		return err
+	}
+	err = cloudClient.SetSecret(cliCtx.Context, path+host+"/AWS_ACCESS_KEY_ID", []byte(app.awsAccessKeyID))
+	if err != nil {
+		return err
+	}
+	err = cloudClient.SetSecret(cliCtx.Context, path+host+"/AWS_SECRET_ACCESS_KEY", []byte(app.awsSecretAccessKey))
+	if err != nil {
+		return err
+	}
+	return nil
+}
+
+func (app *earthlyApp) actionRegistrySetupGCloud(cliCtx *cli.Context, path string, cloudClient *cloud.Client, host string) error {
+	if app.gcpKey == "" {
+		return fmt.Errorf("--gcp-key is missing (or empty)")
+	}
+	err := cloudClient.SetSecret(cliCtx.Context, path+host+"/cred_helper", []byte("gcloud"))
+	if err != nil {
+		return err
+	}
+	return cloudClient.SetSecret(cliCtx.Context, path+host+"/GCP_KEY", []byte(app.gcpKey))
+}
+
+func (app *earthlyApp) actionRegistrySetupUsernamePassword(cliCtx *cli.Context, path string, cloudClient *cloud.Client, host string) error {
+	var err error
 	var password []byte
 	if app.registryPasswordStdin {
 		if app.registryPassword != "" {
@@ -168,6 +256,13 @@ func (app *earthlyApp) actionRegistryLogin(cliCtx *cli.Context) error {
 		return fmt.Errorf("password can not be empty")
 	}
 
+	err = cloudClient.RemoveSecret(cliCtx.Context, path+host+"/cred_helper")
+	if err != nil {
+		if !errors.Is(err, secrets.ErrNotFound) {
+			return err
+		}
+	}
+
 	err = cloudClient.SetSecret(cliCtx.Context, path+host+"/username", []byte(app.registryUsername))
 	if err != nil {
 		return err
@@ -181,12 +276,26 @@ func (app *earthlyApp) actionRegistryLogin(cliCtx *cli.Context) error {
 }
 
 type registryCredentials struct {
-	host     string
-	username string
+	host           string
+	credHelper     string
+	username       string
+	accessID       string
+	gcpClientEmail string
 }
 
-func secretsToRegistryLogins(pathPrefix string, secrets []*cloud.Secret) []*registryCredentials {
-	logins := []*registryCredentials{}
+func extractGCPClientEmail(gcpKey string) (string, error) {
+	data := struct {
+		ClientEmail string `json:"client_email"`
+	}{}
+	err := json.Unmarshal([]byte(gcpKey), &data)
+	if err != nil {
+		return "", err
+	}
+	return data.ClientEmail, nil
+}
+
+func (app *earthlyApp) secretsToRegistryCredentials(pathPrefix string, secrets []*cloud.Secret) []*registryCredentials {
+	credentials := map[string]*registryCredentials{}
 	for _, secret := range secrets {
 		parts := strings.Split(strings.TrimPrefix(secret.Path, pathPrefix), "/")
 		if len(parts) != 2 {
@@ -195,14 +304,39 @@ func secretsToRegistryLogins(pathPrefix string, secrets []*cloud.Secret) []*regi
 		host := parts[0]
 		key := parts[1]
 
-		if key == "username" {
-			logins = append(logins, &registryCredentials{
-				host:     host,
-				username: secret.Value,
-			})
+		rc, ok := credentials[host]
+		if !ok {
+			rc = &registryCredentials{
+				host: host,
+			}
+			credentials[host] = rc
+		}
+
+		switch key {
+		case "cred_helper":
+			rc.credHelper = secret.Value
+		case "username":
+			rc.username = secret.Value
+		case "AWS_ACCESS_KEY_ID":
+			rc.accessID = secret.Value
+		case "GCP_KEY":
+			var err error
+			rc.gcpClientEmail, err = extractGCPClientEmail(secret.Value)
+			if err != nil {
+				app.console.Warnf("Failed to extract client email from %s: %s\n", path.Join(pathPrefix, "GCP_KEY"), err)
+			}
 		}
 	}
-	return logins
+
+	sortedCredentials := []*registryCredentials{}
+	for _, rc := range credentials {
+		sortedCredentials = append(sortedCredentials, rc)
+	}
+	sort.Slice(sortedCredentials, func(i, j int) bool {
+		return sortedCredentials[i].host < sortedCredentials[j].host
+	})
+
+	return sortedCredentials
 }
 
 func (app *earthlyApp) actionRegistryList(cliCtx *cli.Context) error {
@@ -224,16 +358,25 @@ func (app *earthlyApp) actionRegistryList(cliCtx *cli.Context) error {
 	}
 
 	w := tabwriter.NewWriter(os.Stdout, 0, 0, 2, ' ', 0)
-	fmt.Fprintf(w, "%s\t%s\n", "Registry", "username")
-	for _, registry := range secretsToRegistryLogins(path, secrets) {
-		fmt.Fprintf(w, "%s\t%s\n", registry.host, registry.username)
+	fmt.Fprintf(w, "%s\t%s\t%s\n", "Registry", "Cred Helper", "Username/Access ID")
+	for _, rc := range app.secretsToRegistryCredentials(path, secrets) {
+		switch rc.credHelper {
+		case "ecr-login":
+			fmt.Fprintf(w, "%s\t%s\t%s\n", rc.host, rc.credHelper, rc.accessID)
+		case "gcloud":
+			fmt.Fprintf(w, "%s\t%s\t%s\n", rc.host, rc.credHelper, rc.gcpClientEmail)
+		case "":
+			fmt.Fprintf(w, "%s\t%s\t%s\n", rc.host, "none", rc.username)
+		default:
+			fmt.Fprintf(w, "%s\t%s\t%s\n", rc.host, "unknown: "+rc.credHelper, "")
+		}
 	}
 	w.Flush()
 
 	return nil
 }
 
-func (app *earthlyApp) actionRegistryLogout(cliCtx *cli.Context) error {
+func (app *earthlyApp) actionRegistryRemove(cliCtx *cli.Context) error {
 	app.commandName = "registryRemove"
 	path, err := app.getRegistriesPath()
 	if err != nil {
@@ -251,12 +394,17 @@ func (app *earthlyApp) actionRegistryLogout(cliCtx *cli.Context) error {
 	}
 
 	if cliCtx.NArg() > 1 {
-		return fmt.Errorf("only a single registry host can be given")
+		return fmt.Errorf("only a single registry host can be given (or perhaps the host was before the --flags?)")
 	}
 
-	fmt.Printf("Removing login credentials for %s\n", host)
-	for _, secretName := range []string{"username", "password"} {
-		err = cloudClient.RemoveSecret(cliCtx.Context, path+host+"/"+secretName)
+	fmt.Printf("Removing registry credentials for %s\n", host)
+	secrets, err := cloudClient.ListSecrets(cliCtx.Context, path+host)
+	if err != nil {
+		return err
+	}
+
+	for _, secret := range secrets {
+		err = cloudClient.RemoveSecret(cliCtx.Context, secret.Path)
 		if err != nil && !errors.Is(err, cloud.ErrNotFound) {
 			return err
 		}