+++ tests/registry-command/test-gcp.sh
@@ -1,14 +1,14 @@
-#!/usr/bin/env bash
+#!/bin/sh
 set -e
 ./lock.sh
 
-function finish {
+finish() {
   status="$?"
   ./unlock.sh
   if [ "$status" = "0" ]; then
-    echo "test-gcp passed"
+    echo "test-ecr passed"
   else
-    echo "test-gcp failed with $status"
+    echo "test-ecr failed with $status"
   fi
 }
 trap finish EXIT